Switch to pytest for running tests.
The old `numpy.test` only returned True or False whether tests pass or
fail, but did not exit with any different error code. We could change
the code to exit correctly, but using pytest is better because it allows
for skipping or selecting tests in a much simpler way, if necessary.